Pebble/pyproject.toml

[build-system]
requires = ["setuptools>=68","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pebble-raid-fairness"
version = "0.1.2"
description = "Pebble – Raid Time Fairness"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name="Quartz Tools"}]
dependencies = [
  "click>=8.1",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pymongo>=4.7",
  "requests>=2.32",
  "tenacity>=8.5",
  "redis>=5.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "google-api-python-client>=2.141",
  "google-auth>=2.35",
  "google-auth-httplib2>=0.2",
  "google-auth-oauthlib>=1.2",
  "python-dateutil>=2.9",
  "pyyaml>=6.0",
  "mongomock>=4.3"
]

[project.optional-dependencies]
dev = ["pytest>=8.0","pyyaml>=6.0.2","freezegun>=1.5","mypy>=1.11","ruff>=0.5","black>=24.8"]

[project.scripts]
pebble = "pebble.cli:main"

[tool.setuptools.packages.find]
include = ["pebble*"]
exclude = ["docker*","tests*","docs*","scripts*"]
